+++ src/os2datascanner/engine2/tests/test_credit_card_rule.py
@@ -5,26 +5,22 @@ from os2datascanner.engine2.rules import credit_card
 
 @pytest.fixture
 def valid_card():
-    valid_card_number = "4539 1488 0343 6467"
-    return valid_card_number
+    return "4539 1488 0343 6467"
 
 
 @pytest.fixture
 def invalid_card():
-    invalid_card_number = "4111 1111 1111 1112"
-    return invalid_card_number
+    return "4111 1111 1111 1112"
 
 
 @pytest.fixture
 def valid_card_with_dash():
-    valid_card_dash = "4539-1488-0343-6467"
-    return valid_card_dash
+    return "4539-1488-0343-6467"
 
 
 @pytest.fixture
 def invalid_card_with_dash():
-    invalid_card_dash = "4111-1111-1111-1112"
-    return invalid_card_dash
+    return "4111-1111-1111-1112"
 
 
 @pytest.fixture
@@ -34,28 +30,25 @@ def credit_card_rule():
 
 @pytest.fixture
 def four_matches():
-    text_with_matches = """4242424242424242, Lorem ipsum dolor sit amet, 4012888888881881.
+    return """4242424242424242, Lorem ipsum dolor sit amet, 4012888888881881.
     consectetur adipiscing elit. Sed non risus. 4539 1488 0343 6467.
     Suspendisse lectus tortor, dignissim sit amet, adipiscing nec, 5555555555554444 ultricies sed,
     dolor. Cras elementum ultrices diam. Maecenas ligula massa, varius a, semper congue,
     euismod non, mi."""
-    return text_with_matches
 
 
 @pytest.fixture
 def no_matches():
-    text_with_no_matches = """Proin porttitor 4242424242424243, orci nec nonummy molestie,
+    return """Proin porttitor 4242424242424243, orci nec nonummy molestie,
     enim est eleifend mi, non fermentum diam nisl sit amet erat. Duis semper 4012888888881882.
     Duis arcu massa, scelerisque vitae, consequat in, pretium a, enim.
     Pellentesque congue 4111111111111112. Ut in risus volutpat libero pharetra tempor.
     Cras vestibulum bibendum augue 4000056655665557. """
-    return text_with_no_matches
 
 
 @pytest.fixture
 def wrong_number_of_digits():
-    wrong_number_of_digits_text = """4222 2222 2222 2, 3782 822463 10005, 3056 9309 0259 04"""
-    return wrong_number_of_digits_text
+    return "4222 2222 2222 2, 3782 822463 10005, 3056 9309 0259 04"
 
 
 def test_luhn_algorithm_valid_card(valid_card):
@@ -63,11 +56,10 @@ def test_luhn_algorithm_valid_card(valid_card):
     # Arrange
 
     # Act
-    expected_outcome = True
     actual_outcome = credit_card.luhn_algorithm(valid_card)
 
     # Assert
-    assert expected_outcome == actual_outcome
+    assert actual_outcome
 
 
 def test_luhn_algorithm_invalid_card(invalid_card):
@@ -75,11 +67,10 @@ def test_luhn_algorithm_invalid_card(invalid_card):
     # Arrange
 
     # Act
-    expected_outcome = False
     actual_outcome = credit_card.luhn_algorithm(invalid_card)
 
     # Assert
-    assert expected_outcome == actual_outcome
+    assert actual_outcome is False
 
 
 def test_luhn_algorithm_valid_card_with_dash(valid_card_with_dash):
@@ -88,11 +79,10 @@ def test_luhn_algorithm_valid_card_with_dash(valid_card_with_dash):
     # Arrange
 
     # Act
-    expected_outcome = True
     actual_outcome = credit_card.luhn_algorithm(valid_card_with_dash)
 
     # Assert
-    assert expected_outcome == actual_outcome
+    assert actual_outcome
 
 
 def test_luhn_algorithm_invalid_card_with_dash(invalid_card_with_dash):
@@ -101,11 +91,10 @@ def test_luhn_algorithm_invalid_card_with_dash(invalid_card_with_dash):
     # Arrange
 
     # Act
-    expected_outcome = False
     actual_outcome = credit_card.luhn_algorithm(invalid_card_with_dash)
 
     # Assert
-    assert expected_outcome == actual_outcome
+    assert actual_outcome is False
 
 
 def test_match_with_one_match(valid_card, credit_card_rule):